update mainpractice 2 script
cleaned up all the missing values. Per variable was determined if it was a ordinal or categorical variable. Furthermore the mode was often chosen as new value is one was missing
This was done for:
-Miscellanous feautre
- alley
-fence
-Fireplace
-Lot
-Garage
-basement
-Masonry
-MSZoning
-Kitchen
- Sale
Next step will be to make sure that every variable is assigned to the correct type